feat(dx7): wire 6-operator FM engine into Deluge synth pipeline, parse c1.2.0 sessionClips format
Adds Dx7Engine (ChuckUGen subclass) with 32 DX7 algorithm matrices, per-operator
EG state machine, and sine-table lookup. Integrates via new g_dx7_patch bridge
strings — engine creates Dx7Engine instead of MorphingWavetable when a patch is
present for a row.

XML parser now handles the c1.2.0 <sessionClips>/<instrumentClip> format used
by DX7-based songs, with noteDataWithLift parsing and unsigned 32-bit position
fix in DelugeNoteDataMapper. Bridge data must be pushed before engine spork so
init_synth() sees the patch strings.

+++ src/main/java/org/chuck/deluge/xml/DelugeNoteDataMapper.java
@@ -48,8 +48,8 @@ public class DelugeNoteDataMapper {
       String hexLen = data.substring(idx + 8, idx + 16);
       // ignore flags for now
 
-      int pos = Integer.parseInt(hexPos, 16);
-      int len = Integer.parseInt(hexLen, 16);
+      int pos = (int) Long.parseLong(hexPos, 16);
+      int len = (int) Long.parseLong(hexLen, 16);
 
       int step = pos / TICKS_PER_STEP;
       float gate = (float) len / TICKS_PER_STEP;
